implemented QR code generator and moved controllers to package
