+++ sandbox-web/src/main/java/com/exadel/sandbox/team5/controller/ReviewRestController.java
@@ -1,4 +1,4 @@
-package com.exadel.sandbox.team5;
+package com.exadel.sandbox.team5.controller;
 
 import com.exadel.sandbox.team5.entity.Review;
 import com.exadel.sandbox.team5.service.ReviewService;
